requirements.txt
beautifulsoup4==4.14.3
ollama==0.6.1
PyYAML==6.0.3
Requests==2.32.5
streamlit==1.52.2
trafilatura==2.0.0
lxml_html_clean
